Convert rows and cols variables to size_t
There are a number of places where rows and cols being "large"
aren't supported, but at least make the storage for them be
size_t so that serialization is future-proof.

+++ GM/DlgBrdp.h
@@ -76,8 +76,8 @@ public:
     UINT            m_yGridSnapOff;
 
     CellFormType    m_eCellStyle;   // Used to disable cell a cell dimension
-    int             m_nRows;
-    int             m_nCols;
+    size_t          m_nRows;
+    size_t          m_nCols;
     int             m_nCellHt;
     int             m_nCellWd;
     BOOL            m_bStaggerIn;
