+++ GShr/StrLib.h
@@ -32,8 +32,8 @@ char *StrToLong(const char *sp, long *lVal);
 char *SetFileExt(char *fname, const char *ext);
 BOOL StrIsIdentifierChar(int ch);
 BOOL StrIsIdentifier(const char* sp);
-void StrGetAAAFormat(char *szVal, int n);
-void StrLeadZeros(char* szVal, int nWidth);
+void StrGetAAAFormat(char *szVal, size_t n);
+void StrLeadZeros(char* szVal, size_t nWidth);
 void StrTruncatePath(char *pszName);
 void StrExtractFilename(char *pszFName, const char* pszFPath);
 void StrBuildFullFilename(char* pszFull, const char* pszPath,
